# app/swagger.yaml
swagger: "2.0"
info:
  title: em-photographe API documentation
  description: Contains all available API endpoints in this codebase
  version: "1.0.0"
  termsOfService: "http://swagger.io/terms/"
host: localhost:3000
basePath: /
schemes:
  - https
paths:
...
  /articles/{id}:

    # GET /articles/{id} => Read Article
    get:
      tags:
        - Article Module
      summary: Read Article
      description: API to Read an Article
      parameters:
      - name: id
        type: number
        in: path
        description: Id of the Article
        required: true
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/ArticleRes"
        "404":
          description: Not Found
        "500":
          description: Internal Server Error

    # PUT /articles/{id} => Update Article
    put:
      security:
        - Bearer: []
      tags:
        - Article Module
      summary: Update Article
      description: API to Update an Article
      consumes:
        - multipart/form-data
      parameters:
        - name: id
          type: number
          in: path
          description: Id of the Article
          required: true
        - in: body
          name: body
          description: updateArticle() Payload
          required: true
          schema:
            $ref: "#/definitions/ArticleReq"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/BasicRes"
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error

    # DELETE /articles/{id} => Delete Article
    delete:
      security:
        - Bearer: []
      tags:
        - Article Module
      summary: Delete Article
      description: API to Delete an Article
      parameters:
        - name: id
          type: number
          in: path
          description: Id of the Article
          required: true
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: "#/definitions/BasicRes"
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error

  # ! AUTH
  # GET /auth/{id} => Read Avatar
...
  /galleries/{id}:

    # GET /galleries/{id} => Read Gallery
    get:
      tags:
        - Gallery Module
      summary: Read Gallery
      description: API to Read a Gallery
      parameters:
        - name: id
          type: number
          in: path
          description: Id of the Gallery
          required: true
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/GalleryRes"
        "404":
          description: Not Found
        "500":
          description: Internal Server Error

    # PUT /galleries/{id} => Update Gallery
    put:
      security:
        - Bearer: []
      tags:
        - Gallery Module
      summary: Update Gallery
      description: API to Update a Gallery
      consumes:
        - multipart/form-data
      parameters:
        - name: id
          type: number
          in: path
          description: Id of the Gallery
          required: true
        - in: body
          name: body
          description: updateGallery() Payload
          required: true
          schema:
            $ref: "#/definitions/GalleryReq"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/BasicRes"
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error

    # DELETE /galleries/{id} => Delete Article
    delete:
      security:
        - Bearer: []
      tags:
        - Gallery Module
      summary: Delete Gallery
      description: API to Delete a Gallery
      produces:
      - application/json
      parameters:
        - name: id
          type: number
          in: path
          description: Id of the Gallery
          required: true
      responses:
        "204":
          description: No Content
          schema:
            $ref: "#/definitions/BasicRes"
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error

  # ! IMAGES
...
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
definitions:
...
  ArticleReq:
    properties:
      name:
        type: string
        description: article name
      text:
        type: string
        description: article content
      image:
        type: string
        description: article image
      alt:
        type: string
        description: alternative text of article image
      url:
        type: string
        description: article url
      cat:
        type: string
        description: article category

  # AUTH REQUEST
...
  GalleryReq:
    properties:
      name:
        type: string
        description: gallery name
      author:
        type: string
        description: gallery author

  # IMAGE REQUEST
...
  ArticleRes:
    type: object
    properties:
      id:
        type: number
        description: article id
      name:
        type: string
        description: article name
      text:
        type: string
        description: article content
      image:
        type: string
        description: article image
      alt:
        type: string
        description: alternative text of article image
      url:
        type: string
        description: article url
      likes:
        type: string
        description: article likes
      cat:
        type: string
        description: article category
      createdAt:
        type: string
        format: date
        description: article creation
      upatedAt:
        type: string
        format: date
        description: article update

  # AUTH RESPONSE
...
  BasicRes:
    properties:
      status:
        type: integer
        description: status code
      message:
        type: string
        description: confirm message

  # GALLERY RESPONSE
  GalleryRes:
    properties:
      id:
        type: number
        description: gallery id
      name:
        type: string
        description: gallery name
      author:
        type: string
        description: gallery author
      cover:
        type: string
        description: gallery cover image

  # IMAGE RESPONSE
...